requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
python-dotenv
psycopg2-binary